Replace "unsigned char"s and "unsigned short"s with "unsigned"s.

+++ sm.h
@@ -1,7 +1,7 @@
 #include "ll.h"
 #include "item.h"
 
-extern unsigned short counter;
+extern unsigned counter;
 
 struct ll *cll_born(void);
 struct ll *ll_init(struct ll *head, unsigned n, struct ll *(*born)());
