+++ task_can.h
@@ -6,6 +6,6 @@
 
 extern MSG_Q_ID mq_can;
 extern struct ll *current;
-extern unsigned short counter;
+extern unsigned counter;
 
 struct ll *wr_sm(struct ll *p, char *ch, char *data, unsigned n);
